Finalize Python 2/3 issues before migration
In order to get things fully working across the variety
of functions where we use this, I had to make a few changes
that included backing out some recent changes while adding
others (so it wasn't a matter of just rolling back).

TL;DR version:  the reader is always in binary mode, while the
writer operates also in binary mode but will encode strings if
needed.  This makes things work consistently and fixes issues
22 and 23.

Changes:
* Removed decode calls from the reader
* Structure everything so that it handles bytes (so string
literals are encoded as bytes like: `b'\n'`)
* Fix unit tests to do the same
* A few other python 3 fixes
* Unit test that made call to writelines was not correct, should
pass in a list of lines with newlines at the end

With these changes, unit tests work (running tox) and various functions
in Ripley seem to work.